--- tests/abilities/test_more_ability_interactions.py
@@ -1,5 +1,0 @@
-import pytest
-
-from magic_combat import CombatCreature, CombatSimulator, GameState, PlayerState, Color
-from tests.conftest import link_block
-
